Modificando o nome da classe e pacote BasePesonData para BasePerson Modificando campos de BasePerson para simplificar a aplicação
