+++ grails-app/domain/com/quickcharge/baseperson/BasePerson.groovy
@@ -1,0 +1,11 @@
+package com.quickcharge.baseperson
+
+class BasePerson {
+
+    String name
+    String email
+    String cpfCnpj
+    String phone
+    Address address
+
+}
