+++ grails-app/domain/com/quickcharge/baseperson/CEP.groovy
@@ -1,0 +1,5 @@
+package com.quickcharge.baseperson
+
+class CEP {
+    String cep
+}
