+++ grails-app/domain/com/quickcharge/baseperson/CNPJ.groovy
@@ -1,0 +1,7 @@
+package com.quickcharge.baseperson
+
+class CNPJ extends Identifier {
+
+
+
+}
