+++ grails-app/domain/com/quickcharge/baseperson/CPF.groovy
@@ -1,0 +1,7 @@
+package com.quickcharge.baseperson
+
+class CPF extends Identifier {
+
+
+
+}
