+++ grails-app/domain/com/quickcharge/baseperson/Email.groovy
@@ -1,0 +1,6 @@
+package com.quickcharge.baseperson
+
+class Email {
+
+    String email
+}
