--- grails-app/domain/com/quickcharge/personbasedata/BasePersonData.groovy
@@ -1,10 +1,0 @@
-package com.quickcharge.personbasedata
-
-class BasePersonData {
-
-    String name
-    Email email
-    Identifier identifier
-    Address address
-
-}
