--- grails-app/domain/com/quickcharge/personbasedata/CEP.groovy
@@ -1,5 +1,0 @@
-package com.quickcharge.personbasedata
-
-class CEP {
-    String cep
-}
