--- grails-app/domain/com/quickcharge/personbasedata/CNPJ.groovy
@@ -1,7 +1,0 @@
-package com.quickcharge.personbasedata
-
-class CNPJ extends Identifier {
-
-
-
-}
