--- grails-app/domain/com/quickcharge/personbasedata/CPF.groovy
@@ -1,7 +1,0 @@
-package com.quickcharge.personbasedata
-
-class CPF extends Identifier {
-
-
-
-}
