--- grails-app/domain/com/quickcharge/personbasedata/Email.groovy
@@ -1,6 +1,0 @@
-package com.quickcharge.personbasedata
-
-class Email {
-
-    String email
-}
