--- src/test/groovy/com/quickcharge/personbasedata/AddressSpec.groovy
@@ -1,18 +1,0 @@
-package com.quickcharge.personbasedata
-
-import grails.testing.gorm.DomainUnitTest
-import spock.lang.Specification
-
-class AddressSpec extends Specification implements DomainUnitTest<Address> {
-
-    def setup() {
-    }
-
-    def cleanup() {
-    }
-
-    void "test something"() {
-        expect:"fix me"
-            true == false
-    }
-}
