--- src/test/groovy/com/quickcharge/personbasedata/BasePersonDataSpec.groovy
@@ -1,18 +1,0 @@
-package com.quickcharge.personbasedata
-
-import grails.testing.gorm.DomainUnitTest
-import spock.lang.Specification
-
-class BasePersonDataSpec extends Specification implements DomainUnitTest<BasePersonData> {
-
-    def setup() {
-    }
-
-    def cleanup() {
-    }
-
-    void "test something"() {
-        expect:"fix me"
-            true == false
-    }
-}
