--- src/test/groovy/com/quickcharge/personbasedata/CEPSpec.groovy
@@ -1,18 +1,0 @@
-package com.quickcharge.personbasedata
-
-import grails.testing.gorm.DomainUnitTest
-import spock.lang.Specification
-
-class CEPSpec extends Specification implements DomainUnitTest<CEP> {
-
-    def setup() {
-    }
-
-    def cleanup() {
-    }
-
-    void "test something"() {
-        expect:"fix me"
-            true == false
-    }
-}
