--- src/test/groovy/com/quickcharge/personbasedata/CNPJSpec.groovy
@@ -1,18 +1,0 @@
-package com.quickcharge.personbasedata
-
-import grails.testing.gorm.DomainUnitTest
-import spock.lang.Specification
-
-class CNPJSpec extends Specification implements DomainUnitTest<CNPJ> {
-
-    def setup() {
-    }
-
-    def cleanup() {
-    }
-
-    void "test something"() {
-        expect:"fix me"
-            true == false
-    }
-}
