--- src/test/groovy/com/quickcharge/personbasedata/CPFSpec.groovy
@@ -1,18 +1,0 @@
-package com.quickcharge.personbasedata
-
-import grails.testing.gorm.DomainUnitTest
-import spock.lang.Specification
-
-class CPFSpec extends Specification implements DomainUnitTest<CPF> {
-
-    def setup() {
-    }
-
-    def cleanup() {
-    }
-
-    void "test something"() {
-        expect:"fix me"
-            true == false
-    }
-}
