--- src/test/groovy/com/quickcharge/personbasedata/EmailSpec.groovy
@@ -1,18 +1,0 @@
-package com.quickcharge.personbasedata
-
-import grails.testing.gorm.DomainUnitTest
-import spock.lang.Specification
-
-class EmailSpec extends Specification implements DomainUnitTest<Email> {
-
-    def setup() {
-    }
-
-    def cleanup() {
-    }
-
-    void "test something"() {
-        expect:"fix me"
-            true == false
-    }
-}
